RF: rename some methods https://github.com/RedRoverSchool/JenkinsQA_2024_fall/issues/1534
