+++ src/test/java/school/redrover/page/manage/node/NodesConfigPage.java
@@ -6,12 +6,11 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import school.redrover.page.base.BaseConfigPage;
 
-import java.util.List;
 
 public class NodesConfigPage extends BaseConfigPage<NodesConfigPage, NodesProjectPage> {
 
     @FindBy(xpath = "//button[@name='Submit']")
-    private WebElement buttonCreate;
+    private WebElement buttonSave;
 
     public NodesConfigPage(WebDriver driver) {
         super(driver);
@@ -22,13 +21,9 @@ public class NodesConfigPage extends BaseConfigPage<NodesConfigPage, NodesProjec
         return new NodesProjectPage(getDriver());
     }
 
-    public List<String> getNodeList() {
-        return getDriver().findElements(By.xpath("//a[@class='jenkins-table__link model-link inside']")).stream().map(x -> x.getText()).toList();
-    }
+    public NodesPage clickButtonSave() {
+        buttonSave.click();
 
-    public NodesConfigPage clickButtonCreate() {
-        buttonCreate.click();
-
-        return new NodesConfigPage(getDriver());
+        return new NodesPage(getDriver());
     }
 }
