+++ src/test/java/school/redrover/page/manage/node/NodesPage.java
@@ -1,49 +1,29 @@
 package school.redrover.page.manage.node;
 
+import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import school.redrover.page.base.BasePage;
+
+import java.util.List;
 
 public class NodesPage extends BasePage {
 
     @FindBy(xpath = "//a[@href='new']")
     private WebElement buttonNewNode;
 
-    @FindBy(xpath = "//*[@id='name']")
-    private WebElement itemName;
-
-    @FindBy(xpath = "//*[.='Permanent Agent']")
-    private WebElement checkboxAgent;
-
-    @FindBy(xpath = "//button[@name='Submit']")
-    private WebElement buttonCreate;
-
     public NodesPage(WebDriver driver) {
         super(driver);
     }
 
-    public NodesPage clickButtonNewNode() {
+    public NodesProjectPage clickButtonNewNode() {
         buttonNewNode.click();
 
-        return new NodesPage(getDriver());
+        return new NodesProjectPage(getDriver());
     }
 
-    public NodesPage inputName(String name) {
-        itemName.sendKeys(name);
-
-        return this;
-    }
-
-    public NodesPage clickCheckboxAgent() {
-        checkboxAgent.click();
-
-        return this;
-    }
-
-    public NodesConfigPage clickButtonCreate() {
-        buttonCreate.click();
-
-        return new NodesConfigPage(getDriver());
+    public List<String> getNodeList() {
+        return getDriver().findElements(By.xpath("//a[@class='jenkins-table__link model-link inside']")).stream().map(x -> x.getText()).toList();
     }
 }
